[add] mainbackground atom component
